+++ repos/type/global.Double/fromApi.php
@@ -2,5 +2,5 @@
 if (empty ($value))
 	return (float) 0.0;
 
-return (float) Float::validate ($value);
+return (float) Double::validate ($value);
 ?>
